+ QueueController and associate it with CustomerStateMachine
This commit refactors QueueController class that manages the customer queue and associates it with the CustomerStateMachine. When a new customer is spawned, it's added to the customer list and queued up by calling the AddQueue method of the QueueController instance. This change ensures a more natural and orderly customer spawn behavior.