+ CustomerManager: spawn customers with randomly delay and add to queue
This commit adds the ability to spawn customers with random delays in the CustomerManager script. When a new customer is spawned, it is added to the customer list and queued up using the AddQueue method of the QueueController. This change ensures a more natural and less predictable customer spawn behavior, which adds more variety to the game experience.