+ UpgradeButton function and adjustments for Popularity upgrade
In this commit, the UpgradeButton function has been implemented to handle the upgrade functionality.
The code has been adjusted to support the Popularity upgrade, providing additional benefits to the players.
With this update, players can now upgrade the Popularity feature, resulting in a decrease in the Customer Spawn Delay and an increase in the maximum number of customers that can be served.